+++ app/src/main/java/com/mbj/doeat/ui/component/chat/ChatRoomContent.kt
@@ -1,4 +1,4 @@
-package com.mbj.doeat.ui.component
+package com.mbj.doeat.ui.component.chat
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
